python_telegram_bot/requirements.txt
python-telegram-bot==22.5
requests==2.31.0
python-dotenv==1.0.0
